requirements.txt
astroid==2.9.3
autopep8==1.6.0
bidict==0.22.0
click==8.1.3
colorama==0.4.4
ecal5 @ file:///C:/Users/Jonathan/Downloads/ecal5-5.10.1-cp310-cp310-win_amd64.whl
isort==5.10.1
itsdangerous==2.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
numpy==1.22.4
platformdirs==2.5.1
protobuf==3.11.3
pycodestyle==2.8.0
pylint==2.12.2
six==1.16.0
toml==0.10.2
websockets==10.3
wrapt==1.13.3
